+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 For the documentation see the README on the main page.
 
-- now uses BQ 3.0.0-DEV-417 as compatible version
+- now uses BQ 3.0.0-DEV-444 as compatible version to allow `quest-item` tag on `ia` items
     - BQ Dev versions are highly unstable at this point and the AddOn could break any time!